Feat: Added toast messages and filter section is seperated from product page
