+++ package.json
@@ -16,6 +16,7 @@
     "react-responsive": "^9.0.0-beta.6",
     "react-router-dom": "^6.2.2",
     "react-scripts": "4.0.3",
+    "react-toastify": "^8.2.0",
     "uuid": "^8.3.2",
     "web-vitals": "^0.2.4"
   },
