+++ src/styles/base.css
@@ -25,6 +25,10 @@
     --ternary-color: white;
     --ternary-shadow-color: rgb(248, 248, 248);
     --ternary-opaque-color: hsla(0, 0%, 100%, 0.3);
+    --toastify-icon-color-info: var(--primary-color) !important;
+    --toastify-color-progress-info: var(--primary-color) !important;
+    --toastify-color-progress-success: var(--secondary-opaque-color) !important;
+    --toastify-color-progress-error: var(--secondary-opaque-color) !important;
 }
 
 *{
